# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Doc Reader API
  version: "1.0.0"
  description: |
    Single-user personal document server. Paths are root-relative POSIX.
    Document/asset responses use Cache-Control: no-store.
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Health and version
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      summary: Password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Access token issued
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      security: [{ bearerAuth: [] }]
      summary: Current user
      responses:
        "200": { description: OK }
        "401": { description: Unauthorized }
  /auth/pairing-sessions:
    post:
      security: [{ bearerAuth: [] }]
      summary: Create 60s one-time pairing session
      responses:
        "201": { description: Pairing session created }
  /auth/pairing/exchange:
    post:
      summary: Exchange pairing payload for access token
      responses:
        "200": { description: Token issued }
        "401": { description: Invalid/expired/consumed }
  /auth/logout:
    post:
      security: [{ bearerAuth: [] }]
      summary: Revoke current token
      responses:
        "200": { description: OK }
  /bootstrap:
    get:
      security: [{ bearerAuth: [] }]
      summary: User, roots, features
      responses:
        "200": { description: OK }
  /home:
    get:
      security: [{ bearerAuth: [] }]
      summary: Pinned and recent documents
      responses:
        "200": { description: OK }
  /documents/pin:
    put:
      security: [{ bearerAuth: [] }]
      summary: Pin or unpin a document
      responses:
        "200": { description: OK }
  /documents/opened:
    post:
      security: [{ bearerAuth: [] }]
      summary: Update recent order
      responses:
        "200": { description: OK }
  /search:
    get:
      security: [{ bearerAuth: [] }]
      summary: FTS search
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200": { description: Search hits without full content }
  /tree:
    get:
      security: [{ bearerAuth: [] }]
      summary: Lazy tree children
      parameters:
        - in: query
          name: root_id
          required: true
          schema: { type: string }
        - in: query
          name: path
          schema: { type: string }
      responses:
        "200": { description: Direct children }
  /documents:
    get:
      security: [{ bearerAuth: [] }]
      summary: Read document (raw_content + revision)
      parameters:
        - in: query
          name: root_id
          required: true
          schema: { type: string }
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Document
          headers:
            Cache-Control:
              schema: { type: string, example: no-store }
    post:
      security: [{ bearerAuth: [] }]
      summary: Create markdown/txt/json
      responses:
        "201": { description: Created }
        "409": { description: Path exists }
    put:
      security: [{ bearerAuth: [] }]
      summary: Update with if_match_revision
      responses:
        "200": { description: Updated }
        "409": { description: Conflict with remote document body }
  /directories:
    post:
      security: [{ bearerAuth: [] }]
      summary: Create directory
      responses:
        "201": { description: Created }
  /entries/move:
    patch:
      security: [{ bearerAuth: [] }]
      summary: Move or rename
      responses:
        "200": { description: Moved }
        "409": { description: Target exists }
  /entries:
    delete:
      security: [{ bearerAuth: [] }]
      summary: Soft-delete into trash
      responses:
        "200": { description: Trash record returned }
  /trash:
    get:
      security: [{ bearerAuth: [] }]
      summary: List recoverable trash
      responses:
        "200": { description: OK }
  /trash/{id}/restore:
    post:
      security: [{ bearerAuth: [] }]
      summary: Restore trash entry
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Restored }
        "409": { description: Target exists }
  /trash/{id}:
    delete:
      security: [{ bearerAuth: [] }]
      summary: Permanent delete (confirm required)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Deleted }
  /assets:
    get:
      security: [{ bearerAuth: [] }]
      summary: Stream image asset
      parameters:
        - in: query
          name: root_id
          required: true
          schema: { type: string }
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Image bytes
          headers:
            Cache-Control:
              schema: { type: string, example: no-store }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT